Remove Sentinel integration with Loggie and Collar
We deprecated Loggie and replaced it with Pulse, so we had to update
Sentinel according to this.

However, we thought that neither Pulse, Loggie or Collar belong to
Sentinel. It is trivial to implement these integrations in client code
base and we avoid issues such as this one or with versioning.